server/requirements.txt
dashscope==1.23.2
Flask>=3.0,<4
Werkzeug>=3.0,<4
Flask-Cors>=4,<5
Flask-SocketIO>=5.3.6,<6
Flask-Limiter>=3.5,<4
flask-smorest>=0.45.0
pydantic>=2,<3
redis==6.0.0
requests==2.32.3
websocket-client==1.8.0
gevent>=23.9.1
gevent-websocket>=0.10.1
bleak>=0.22.3
APScheduler==3.10.4
upnpclient>=0.0.8
ssdpy>=0.4.0
flask_sqlalchemy==3.1.1
miservice_fork>=0.1.0
python-dotenv>=1.0.0
nest_asyncio>=1.5.6
pikepdf>=8.0.0
lxml>=5.0.0
dbus-python>=1.2.18; platform_system == "Linux"
pytest==8.3.4
pytest-cov
fakeredis